+++ apps/ethereum_jsonrpc/lib/ethereum_jsonrpc/parity/fetched_beneficiaries.ex
@@ -180,6 +180,16 @@ defmodule EthereumJSONRPC.Parity.FetchedBeneficiaries do
   defp get_address_type(reward_type, index) when reward_type == "external" and index == 8, do: :validator
   defp get_address_type(reward_type, index) when reward_type == "external" and index == 9, do: :validator
   defp get_address_type(reward_type, index) when reward_type == "external" and index == 10, do: :validator
+  defp get_address_type(reward_type, index) when reward_type == "external" and index == 11, do: :validator
+  defp get_address_type(reward_type, index) when reward_type == "external" and index == 12, do: :validator
+  defp get_address_type(reward_type, index) when reward_type == "external" and index == 13, do: :validator
+  defp get_address_type(reward_type, index) when reward_type == "external" and index == 14, do: :validator
+  defp get_address_type(reward_type, index) when reward_type == "external" and index == 15, do: :validator
+  defp get_address_type(reward_type, index) when reward_type == "external" and index == 16, do: :validator
+  defp get_address_type(reward_type, index) when reward_type == "external" and index == 17, do: :validator
+  defp get_address_type(reward_type, index) when reward_type == "external" and index == 18, do: :validator
+  defp get_address_type(reward_type, index) when reward_type == "external" and index == 19, do: :validator
+  defp get_address_type(reward_type, index) when reward_type == "external" and index == 20, do: :validator
   defp get_address_type(reward_type, _index) when reward_type == "block", do: :validator
   defp get_address_type(reward_type, _index) when reward_type == "uncle", do: :uncle
   defp get_address_type(reward_type, _index) when reward_type == "emptyStep", do: :validator